exception-closure2.js

+++ 6.exception-closure/task.js
@@ -1,5 +1,5 @@
 ﻿function parseCount(value) {
-	const parsed = Number.parseInt(value);
+	const parsed = Number.parseFloat(value);
 	if (isNaN(parsed)) {
 		throw new Error("Невалидное значение");
 	}
